pyproject.toml
[project]
name = "database-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.16.5",
    "cryptography>=46.0.2",
    "fastapi>=0.118.2",
    "langchain-google-genai>=2.1.12",
    "langchain-openai>=0.3.35",
    "langgraph>=0.6.8",
    "psycopg2-binary>=2.9.10",
    "pymysql>=1.1.2",
    "python-dotenv>=1.1.1",
    "sqlalchemy>=2.0.43",
    "uvicorn>=0.37.0",
]
